feat: implement culvert connection serialization and add formatting improvements
- Add culvert parsing support to connection parser by importing parseCulvertData
- Implement culvert serialization in connection serializer using serializeCulvertGroups
- Add extra newlines after specific keys as requested:
  * 2D Locked: add blank line in storageAreaSerializer
  * Conn Culv Bottom: add blank line in culvertSerializer
  * Conn HTab HWMax: add blank line in connectionSerializer
- Fix connection name padding to use 16-character fixed width formatting
- Fix Conn Weir WD parsing to use parseFloat instead of parseInt for decimal support
- Add round-trip test infrastructure with line-by-line comparison script
- Update comparison script formatting for better readability

+++ src/parsers/geometry/connectionParser.ts
@@ -82,7 +82,7 @@ export function parseConnectionData(
       connection.overflowMethod2D = parseKeyValue(currentLine).value.trim().toLowerCase() === "true"
       index++
     } else if (currentLine.includes("Conn Weir WD=")) {
-      connection.weirWD = parseInt(parseKeyValue(currentLine).value.trim())
+      connection.weirWD = parseFloat(parseKeyValue(currentLine).value.trim())
       index++
     } else if (currentLine.includes("Conn Weir Coef=")) {
       connection.weirCoefficient = parseFloat(parseKeyValue(currentLine).value.trim())
